+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 #CHANGELOG
 All notable changes to this project will be documented in this file.
 
+##0.0.2 - 28.11.2019
+
+###Added
+- aliases for modules
+
+###Patched
+- closed access to project settings from get request
+
 ##0.0.1 - 25.11.2019
 
 early first dev version